separated logic to a new file
